requirements.txt
pytesseract
youtube-dl
youtube-search-python
pillow